requirements.txt
PySide6==6.6.0
Pillow==10.1.0
openpyxl==3.1.2
pyserial==3.5
